Initial code migration to split out `tortoise` from core
