+++ sanic_security/authentication.py
@@ -8,12 +8,13 @@ from argon2.exceptions import VerifyMismatchError
 from sanic import Sanic
 from sanic.log import logger
 from sanic.request import Request
-from tortoise.exceptions import IntegrityError, DoesNotExist
 
 from sanic_security.configuration import config as security_config
 from sanic_security.exceptions import (
     NotFoundError,
     CredentialsError,
+    DoesNotExist,
+    IntegrityError,
     SessionError,
     DeactivatedError,
     ExpiredError,
